Converted most similarity values from double to float.

+++ src/main/java/org/scify/jedai/blockprocessing/comparisoncleaning/CardinalityNodePruning.java
@@ -130,10 +130,10 @@ public class CardinalityNodePruning extends CardinalityEdgePruning {
         }
 
         topKEdges.clear();
-        minimumWeight = Double.MIN_VALUE;
+        minimumWeight = Float.MIN_VALUE;
         for (TIntIterator iterator = validEntities.iterator(); iterator.hasNext();) {
             int neighborId = iterator.next();
-            double weight = getWeight(entityId, neighborId);
+            float weight = getWeight(entityId, neighborId);
             if (!(weight < minimumWeight)) {
                 final Comparison comparison = new Comparison(cleanCleanER, -1, neighborId);
                 comparison.setUtilityMeasure(weight);
